+++ packages/page-spy-harmony/library/src/main/ets/types/lib/system.d.ts
@@ -1,17 +1,17 @@
-import { Browser, OS } from './device';
-
 export type Category =
   | 'Element'
   | 'Javascript'
   | 'Feature'
   | 'Network'
   | 'Storage';
+
 export interface FeatureDescriptor {
   title: string;
   keyPath?: string;
   customTest?: string;
   supported?: boolean;
 }
+
 export type Feature = Record<Category, Record<string, FeatureDescriptor>>;
 
 export interface DataItem {